+++ ios/Podfile
@@ -8,7 +8,9 @@ require Pod::Executable.execute_command('node', ['-p',
 platform :ios, min_ios_version_supported
 prepare_react_native_project!
 
-linkage = ENV['USE_FRAMEWORKS']
+use_modular_headers!
+
+linkage = ENV['USE_FRAMEWORKS'] || 'static' # 기본값 static
 if linkage != nil
   Pod::UI.puts "Configuring Pod with #{linkage}ally linked Frameworks".green
   use_frameworks! :linkage => linkage.to_sym
