+++ src/services/navigation/AppNavigator.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import TabNavigator from './TabNavigator';
+import TabNavigator from './TabNavigator.tsx';
 
 export default function AppNavigator() {
   return (
